requirements.txt
streamlit>=1.28.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
